Move struct dtype test
